requirements.txt
python-dotenv==1.1.1
python-telegram-bot==22.5
isodate==0.7.2
apscheduler==3.11.0
glocaltokens==0.7.6
python-dateutil==2.9.0.post0
pydantic==2.12.3
pytz==2025.2
requests==2.32.5
tzlocal==5.3.1
google-cloud-pubsub==2.26.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1